Moved all alert handling to alerts service.
